requirements.txt
torch >= 2.1.0
numpy >= 1.18.1
kaggle >= 1.6.0
transformers >= 4.36.2
pandas >= 2.1.0
pytorch-lightning >= 2.1.0
hydra-core >= 1.3.2
wandb >= 0.16.0
pytest >= 7.4.0
